front end and data refactor

+++ src/Resources/ExportReturnDetailResource.cs
@@ -28,12 +28,12 @@
 
         public int? NumCartons { get; set; }
 
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
 
-        public int? Width { get; set; }
+        public double? Width { get; set; }
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
-        public int? Depth { get; set; }
+        public double? Depth { get; set; }
     }
 }
